requirements.txt
unidecode
pyrogram
TgCrypto
requests
pytz
git+https://github.com/ytdl-org/youtube-dl
youtube_search
shazamio
telegraph
pillow
git+https://github.com/ali-mmagneto/gerek02
git+https://github.com/ali-mmagneto/gerek01
asyncio
pytube
PyMovieDb 
hachoir
ffmpeg-python
bs4
lxml
youtube-video-scraping
